Breakout DrawablePath
