+++ src/Svg.Skia/Svg/SvgExtensions.cs
@@ -5140,57 +5140,6 @@ namespace Svg.Skia
         void Draw(SKCanvas canvas, Attributes ignoreAttributes, DrawableBase? until);
     }
 
-    internal abstract class DrawablePath : DrawableBase
-    {
-        public SKPath? Path;
-
-        public List<DrawableBase>? MarkerDrawables;
-
-        protected DrawablePath()
-            : base()
-        {
-        }
-
-        public override void OnDraw(SKCanvas canvas, Attributes ignoreAttributes, DrawableBase? until)
-        {
-            if (until != null && this == until)
-            {
-                return;
-            }
-
-            if (Fill != null && Path != null)
-            {
-                canvas.DrawPath(Path, Fill);
-            }
-
-            if (Stroke != null && Path != null)
-            {
-                canvas.DrawPath(Path, Stroke);
-            }
-
-            if (MarkerDrawables != null)
-            {
-                foreach (var drawable in MarkerDrawables)
-                {
-                    drawable.Draw(canvas, ignoreAttributes, until);
-                }
-            }
-        }
-
-        public override void PostProcess()
-        {
-            base.PostProcess();
-
-            if (MarkerDrawables != null)
-            {
-                foreach (var drawable in MarkerDrawables)
-                {
-                    drawable.PostProcess();
-                }
-            }
-        }
-    }
-
     internal abstract class DrawableContainer : DrawableBase
     {
         public readonly List<DrawableBase> ChildrenDrawables;
